Don't automatically fill in common start if it doesn't match.
This avoids the loss of characters when tab-completing something where
two possibilities share a common start that doesn't match all of the
chars in the searched string.

For example: @dsu could expand to @deadsuperhero or @darwinsurvivor. The
only common start between these is @d, so tab-completing with automatic
common start would lose "su" for no good reason. This change prevents
that happening. (Long commit message? Tough.)