Fixed a bug involving getting the length of None
I had an incident in which IdentiCurse was trying to find the length
of an object of type NoneType, and raising an exception.  A three-
line patch fixes the problem, but I fear the root of the issue may
be in StatusNet itself.  Not certain, however, so don't take my word
for it.